requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-huggingface
faiss-cpu
python-dotenv
pypdf
sentence-transformers

